+++ client/src/App.jsx
@@ -1,0 +1,11 @@
+
+import React from 'react'
+import Dashboard from './Components/Dashboard/Dashboard'
+
+const App = () => {
+  return (
+    <Dashboard />
+  )
+}
+
+export default App
