+++ client/src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReaactDOM from 'react-dom';
-import App from './components/App/App'
+import App from './App.jsx'
 import 'mdb-react-ui-kit/dist/css/mdb.min.css';
 import "@fortawesome/fontawesome-free/css/all.min.css";
+import './index.css';
 
 ReaactDOM.render(<App />, document.getElementById('root'));
